Module cocktail::rle

A set of function to convert DNA sequence in Run Length Encoding (RLE) and work with it. RLE is a small compression method where repetition of lettre are replace by the number of repetition and the lettre

Exemple: ACCTGGGAAT -> 1A2CT3G2A1T

A RLE symbol are encode on a byte. The two lowweight bits encode the nucletoide value, see kmer module for more details.

A or a -> 00

C or c -> 01

T or t -> 10

G or g -> 11

The six heavyweight bits encode the number of repetition plus one

Functions

rle2bit

Convert a rle in 2bit representation repetition is ignore.

rle2kmer

Convert a rle sequence in 2 bit representation repetition is ignore. If sequence is larger than 32 only the last 32 nuc is store.

rle2seq

Convert a rle in String.

seq2rle

Convert a sequence in rle representation