Module cocktail::kmer

A set of function to convert small sequence (less than 32 nucleotide) in 2 bit representation.

A or a -> 00

C or c -> 01

T or t -> 10

G or g -> 11

We use the second and thrid bit of each value provide, if you provide no ACTG value this function silently convert to A, C, T or G, for exemple N or n is convert in G.

With this coding and if kmer size is odd, if the popcount of forward is odd the popcount of reverse is even. In this library if a kmer have even popcount is the canonical kmer.

If we work only with canonical kmer, we can remove one bit at any extremity. To reconstruct lost bit, if result have even popcount we add a 0, if it’s ood we add 1.

This 2bit coding is inspired by https://cs.stackexchange.com/questions/82644/compact-mapping-from-an-involuted-set

Functions

bit2nuc

Convert the 2bit representation of a nucleotide in nucleotide

canonical

Take a kmer and return the canonical form

comp

Return the complement of kmer

get_first_bit

Return true if the right bit of kmer is 1

get_hash_space_size

Return the cardinality of canonical hash set for a given kmer size

get_kmer_space_size

Return the cardinality of canonical kmer set for a given kmer size

hash

Take a subseq and return the canonical kmer with out the rightest bit

kmer2seq

Convert a 2 bit repersentation in String.

nuc2bit

Convert a nucleotide in 2bit representation, by use conversion present in seq2bit

parity_even

Return true if the kmer parity is even

remove_first_bit

Return the kmer without the rightest bit of kmer

rev

Return the reverse of kmer

revcomp

Return the reverse complement of kmer

seq2bit

Convert a sequence in 2 bit representation if suseq is larger than 32 only the last 32 nuc is store